updated iplcorestats

+++ src/main/java/com/careerit/jfs/iplcorestats/TeamAmountRecord.java
@@ -1,4 +1,4 @@
 package com.careerit.jfs.iplcorestats;
 
-public record TeamAmountRecord(String teamName, double amount) {
+public record TeamAmountRecord(String teamName, int amount) {
 }
